reorganized test drivers, updated imports accordingly, included test  properties method in stanford830, and added test file for pytest to automatically test the drivers_test_unit.py is working as expected.
