+++ pyscan/drivers/__init__.py
@@ -77,6 +77,8 @@ except ModuleNotFoundError:
 from .new_instrument import new_instrument
 
 # Test Devices
-from .test_voltage import TestVoltage
+from .test_drivers.test_voltage import TestVoltage
 
-from .test_instrument_driver import TestInstrumentDriver
+from .test_drivers.test_instrument_driver import TestInstrumentDriver
+
+from .test_drivers.drivers_test_unit import test_driver, check_has_attributes, check_attribute_values
