+++ pyscan/drivers/stanford830.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 from .instrument_driver import InstrumentDriver
+from .test_drivers.drivers_test_unit import test_driver
 import numpy as np
 from time import sleep
 
@@ -224,6 +225,9 @@ class Stanford830(InstrumentDriver):
             'return_type': int})
 
         self.update_properties()
+
+    def test_properties(self):
+        test_driver(self)
 
     def update_properties(self):
         # Reference and phase
