+++ test/drivers/test_drivers_test_unit.py
@@ -1,0 +1,16 @@
+from pyscan.drivers.test_drivers.test_instrument_driver import TestInstrumentDriver
+from pyscan.drivers.test_drivers.test_voltage import TestVoltage
+from pyscan.drivers.test_drivers.drivers_test_unit import test_driver
+
+
+def test_drivers_test_unit():
+    test_instrument = TestInstrumentDriver()
+    test_driver(test_instrument)
+
+    test_voltage = TestVoltage()
+
+    # query functions for test voltage have been bypassed to decouple dependencies so it remains modifiable
+    test_driver(test_voltage)
+
+
+test_drivers_test_unit()
