+++ test/drivers/test_test_voltage.py
@@ -2,10 +2,8 @@
 Pytest functions to test the Runinfo class
 '''
 
-import pyscan as ps
+from pyscan.drivers.test_drivers.test_voltage import TestVoltage
 import pytest
-import importlib
-importlib.reload(ps)
 
 
 def test_testvoltage():
@@ -18,7 +16,7 @@ def test_testvoltage():
     """
 
     # set up v1 as representative for testing
-    v1 = ps.TestVoltage()
+    v1 = TestVoltage()
 
     # ########## add 2 more test voltages for testing...
 
@@ -86,3 +84,6 @@ def test_testvoltage():
     v1.output_state = 'off'
     assert v1._output_state == 'off', "TestVoltage _output_state does not get machine value after 'off' key user input"
     assert v1.output_state == 'off', "TestVoltage output_state does not get machine value after 'off' key user input"
+
+
+test_testvoltage()
